requirements.txt
asgiref            3.8.1
boto3              1.34.79
botocore           1.34.79
certifi            2024.2.2
charset-normalizer 3.3.2
Django             4.2.11
django-ckeditor    6.7.1
django-jalali      6.2.0
django-js-asset    2.2.0
django-storages    1.14.2
idna               3.6
jalali_core        1.0.0
jdatetime          5.0.0
jmespath           1.0.1
kavenegar          1.1.2
pip                24.0
psycopg2           2.9.9
python-dateutil    2.9.0.post0
requests           2.31.0
s3transfer         0.10.1
setuptools         68.2.0
six                1.16.0
sqlparse           0.4.4
typing_extensions  4.10.0
urllib3            1.26.18
wheel              0.41.2
